feature(contracts): ✨ extract experimental capability interfaces
Introduce composable capability interfaces under `Darsyn\IP\Contracts\`.
Tagged `@experimental` ahead of the 7.0 interface decomposition.
