+++ tests/Version/IPv4Test.php
@@ -4,6 +4,13 @@ declare(strict_types=1);
 
 namespace Darsyn\IP\Tests\Version;
 
+use Darsyn\IP\Contracts\ArithmeticInterface;
+use Darsyn\IP\Contracts\Classification4Interface;
+use Darsyn\IP\Contracts\ClassificationInterface;
+use Darsyn\IP\Contracts\ComparisonInterface;
+use Darsyn\IP\Contracts\Output4Interface;
+use Darsyn\IP\Contracts\OutputInterface;
+use Darsyn\IP\Contracts\VersionIdentityInterface;
 use Darsyn\IP\Exception\InvalidCidrException;
 use Darsyn\IP\Exception\InvalidIpAddressException;
 use Darsyn\IP\Exception\WrongVersionException;
@@ -25,6 +32,20 @@ class IPv4Test extends TestCase
     public function resetProtocolFormatter(): void
     {
         IP::setProtocolFormatter(new ConsistentFormatter());
+    }
+
+    /** @test */
+    #[PHPUnit\Test]
+    public function testImplementsCapabilityInterfaces(): void
+    {
+        $ip = IP::factory('127.0.0.1');
+        $this->assertInstanceOf(VersionIdentityInterface::class, $ip);
+        $this->assertInstanceOf(ComparisonInterface::class, $ip);
+        $this->assertInstanceOf(ArithmeticInterface::class, $ip);
+        $this->assertInstanceOf(OutputInterface::class, $ip);
+        $this->assertInstanceOf(Output4Interface::class, $ip);
+        $this->assertInstanceOf(ClassificationInterface::class, $ip);
+        $this->assertInstanceOf(Classification4Interface::class, $ip);
     }
 
     /**
